Improvements (#47)

+++ test/utils/GrantFundTestHelper.sol
@@ -92,7 +92,7 @@ abstract contract GrantFundTestHelper is Test {
         return false;
     }
 
-    function _claimDelegateReward(GrantFund grantFund_, address voter_, uint256 distributionId_, uint256 claimedReward_) internal {
+    function _claimDelegateReward(GrantFund grantFund_, address voter_, uint24 distributionId_, uint256 claimedReward_) internal {
         changePrank(voter_);
         vm.expectEmit(true, true, false, true);
         emit DelegateRewardClaimed(voter_, distributionId_, claimedReward_);
@@ -143,7 +143,7 @@ abstract contract GrantFundTestHelper is Test {
         uint256 expectedProposalId = grantFund_.hashProposal(targets_, values_, proposalCalldatas_, keccak256(bytes(description)));
         uint256 startBlock = block.number.toUint64() + grantFund_.votingDelay().toUint64();
 
-        (, , , uint256 endBlock, , ) = grantFund_.getDistributionPeriodInfo(grantFund_.getDistributionId());
+        (, , uint48 endBlock, , , ) = grantFund_.getDistributionPeriodInfo(grantFund_.getDistributionId());
 
         // submit proposal
         changePrank(proposer_);
